debug

+++ hotspot/src/share/vm/interpreter/interpreterRuntime.cpp
@@ -154,10 +154,10 @@ IRT_ENTRY(void, InterpreterRuntime::resolve_ldc(JavaThread* thread, Bytecodes::C
 IRT_END
 
 IRT_ENTRY(void, InterpreterRuntime::_print(JavaThread* thread, oopDesc* obj))
-  uint64_t objOffset = (uint64_t)obj - Universe::getHeapStart();
+  uint64_t objOffset = (uint64_t)obj - (uint64_t)Universe::getHeapStart();
 
   uint64_t regionI = objOffset /(_R_SIZE);
-  uint64_t position = regionI + Universe::getRegionTable();
+  uint64_t position = regionI + (uint64_t)Universe::getRegionTable();
   printf("%p, %p, %p, %p, %p, %p\n", obj, Universe::getHeapStart(), objOffset, regionI,Universe::getRegionTable(), position);
   fflush(stdout);
   if (*(int *)position > 0){
